Make Meta struct public
