+++ buildpacks/ruby/src/layers/shared.rs
@@ -1,5 +1,6 @@
 use cache_diff::CacheDiff;
 use commons::display::SentenceList;
+pub(crate) use commons::layer::cache_buddy::Meta;
 use libcnb::build::BuildContext;
 use libcnb::data::layer::LayerName;
 use libcnb::layer::{CachedLayerDefinition, InvalidMetadataAction, LayerRef, RestoredLayerAction};
@@ -93,46 +94,6 @@ where
                 "Clearing cache due to invalid metadata serialization error: {error}"
             )),
         ),
-    }
-}
-
-/// Either contains metadata or a message describing the state
-///
-/// Why: The `CachedLayerDefinition` allows returning information about the cache state
-/// from either `invalid_metadata_action` or `restored_layer_action` functions.
-///
-/// Because the function returns only a single type, that type must be the same for
-/// all possible cache conditions (cleared or retained). Therefore, the type must be
-/// able to represent information about the cache state when it's cleared or not.
-///
-/// This struct implements `Display` and `AsRef<str>` so if the end user only
-/// wants to advertise the cache state, they can do so by passing the whole struct
-/// to `format!` or `println!` without any further maniuplation. If they need
-/// to inspect the previous metadata they can match on the enum and extract
-/// what they need.
-///
-/// - Will only ever contain metadata when the cache is retained.
-/// - Will contain a message when the cache is cleared, describing why it was cleared.
-///   It is also allowable to return a message when the cache is retained, and the
-///   message describes the state of the cache. (i.e. because a message is returned
-///   does not guarantee the cache was cleared).
-pub(crate) enum Meta<M> {
-    Message(String),
-    Data(M),
-}
-
-impl<M> std::fmt::Display for Meta<M> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.as_ref())
-    }
-}
-
-impl<M> AsRef<str> for Meta<M> {
-    fn as_ref(&self) -> &str {
-        match self {
-            Meta::Message(s) => s.as_str(),
-            Meta::Data(_) => "Using cache",
-        }
     }
 }
